Refactor stdlib directory structures
