+++ include/signal.h
@@ -17,6 +17,7 @@
 #define SIGHUP    1
 #define SIGINT    2
 #define SIGQUIT   3
+#define SIGABRT   6
 #define SIGBUS    7
 #define SIGFPE    8
 #define SIGKILL   9
