+++ src/misc/assert.c
@@ -6,6 +6,7 @@
  * Change Logs:
  * Date           Author       Notes
  * 2023/7/03      bernard      first version
+ * 2025/2/1      Chris-godz    the second version
  */
 
 #include <stdio.h>
@@ -15,5 +16,6 @@
 mlibc_noreturn void __assert_fail(const char* expr, const char* file, int line)
 {
     fprintf(stderr, "%s:%d: Assertion failed: %s\n", file, line, expr);
-    abort(); 
+    abort();
+    while(1);
 }
