+++ src/stdlib/exit.c
@@ -5,13 +5,23 @@
  *
  * Change Logs:
  * Date           Author       Notes
- * 2024/8/26  0Bitbiscuits the first version
+ * 2024/6/4   0Bitbiscuits the first version
+ * 2025/2/1   Chris-godz   the second version
  */
 #include <stdlib.h>
 #include <compiler.h>
+#include <unistd.h>
+
+#define NATS	32
+void (*_Atfuns[NATS])(void) = {0};
+size_t _Atcount = {NATS};
 
 mlibc_weak void exit(int status)
 {
+    while (_Atcount < NATS)
+    {
+        (*_Atfuns[_Atcount++])();
+    }
     fflush(NULL);
-    abort();
+    _exit(status);
 }
